+++ gulpfile.js
@@ -3,7 +3,6 @@ const cfg = require('./paths.json');
 
 const {cleanAll} = require('./build/clean-task');
 const {tscBuild} = require('./build/tsc-task');
-const {tarBuild} = require('./build/tar-task');
 const {lessCopy, lessBuildProd} = require('./build/less-task');
 const {catLog} = require('./build/common');
 const {lintStyle, lintTypeScript} = require('./build/linting-task');
@@ -14,10 +13,8 @@ const buildPolyfills = tscBuild({src: cfg.polyfills.ts, base: cfg.polyfills.base
 const buildLess = lessCopy({src: cfg.src.less, base: cfg.src.base}, cfg.src.dest);
 const buildCss = lessBuildProd({src: cfg.src.css, base: cfg.src.base}, cfg.src.dest);
 
-const clean = cleanAll([...cfg.src.destPaths, ...cfg.polyfills.destPaths, ...cfg.tar.destPaths]);
+const clean = cleanAll([...cfg.src.destPaths, ...cfg.polyfills.destPaths]);
 const build = gulp.series(clean, gulp.parallel(buildModules, buildPolyfills, buildLess, buildCss));
-
-const tar = gulp.series(build, tarBuild(cfg.tar.dest));
 
 // === LINTER TASKS ===
 const lintTS = lintTypeScript(cfg.lint.ts);
@@ -25,7 +22,6 @@ const lintCSS = lintStyle(cfg.lint.less)
 const lint = gulp.series(gulp.parallel(lintTS, lintCSS), catLog('Linting passed'));
 
 module.exports = {
-  tar,
   lint,
   lintTS,
   lintCSS,