+++ src/modules/all.ts
@@ -5,11 +5,6 @@ export * from './esl-traversing-query/core';
 
 export * from './esl-image/core';
 
-import './esl-media/providers/iframe-provider';
-import './esl-media/providers/html5/audio-provider';
-import './esl-media/providers/html5/video-provider';
-import './esl-media/providers/youtube-provider';
-import './esl-media/providers/brightcove-provider';
 export * from './esl-media/core';
 
 export * from './esl-popup/core';
@@ -21,3 +16,6 @@ export * from './esl-tab/core';
 export * from './esl-scrollable-tabs/core';
 
 export * from './esl-scrollbar/core';
+
+export * from './esl-alert/core';
+