+++ src/modules/esl-base-popup/core/esl-base-popup.ts
@@ -98,7 +98,7 @@ export class ESLBasePopup extends ESLBaseElement {
   protected bindBodyClickTracking(track: boolean) {
     document.body.removeEventListener('click', this._onBodyClick);
     if (track) {
-      document.body.addEventListener('click', this._onBodyClick);
+      document.body.addEventListener('click', this._onBodyClick, true);
     }
   }
   protected bindHoverStateTracking(track: boolean) {