+++ src/modules/esl-utils/fixes/ie-fixes.ts
@@ -1,0 +1,7 @@
+/** Fix IE browser to allow to display alert under iframe */
+export function createIframe(): HTMLElement {
+  const iframe = document.createElement('iframe');
+  iframe.className = 'ie-zindex-fix';
+  iframe.src = 'about:blank';
+  return iframe;
+}